app/requirements.txt
fastapi
fastapi_mail==1.2.0
pydantic==1.10.2
python-decouple==3.6
pytz==2022.5
SQLAlchemy==1.4.41
uvicorn==0.19.0
psycopg2-binary==2.9.5
